增加前端页面展示

+++ frontend/model/page.go
@@ -3,7 +3,7 @@ package model
 import "crawler/concurrent/engine"
 
 type SearchResult struct {
-	Hits     int64
+	Hits     float64
 	Start    int
 	Query    string
 	PrevFrom int
